Fix a potential bug in the sum of three range
some of three would count the last depths sums even though there wrere not three.

These sums were not larger than the porevious sum  so they did not affect the answer.

This fix is more accurate though as it ommits the last deoths sums that do not have three numbers to sum.